% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeihnMatch.R
\name{FeihnMatch}
\alias{FeihnMatch}
\title{Takes a metabolite info table from the ConsensusAlign output and annotates with the Feihn library}
\usage{
FeihnMatch(PeakInfoTable, FeihnLibrary, numCores = 4, commonIons = c(),
  RT1_Standards = NULL, RT1Penalty = 1)
}
\arguments{
\item{PeakInfoTable}{Metabolite info table (the second data frame in the output from ConsensusAlign)}

\item{FeihnLibrary}{Provide Feihn library generated from MakeReference function to ID metabolites with retention index. (Use data(FeihnStandards_030617) to find)}

\item{numCores}{Number of cores used to parallelize alignment. See parallel package. Defaults to 4.}

\item{commonIons}{Provide a vector of ions to ignore from the FindProblemIons function. Defaults to empty vector.}

\item{RT1_Standards}{Vector of standard names used to adjust first retention time. All names must be found in input files. Defaults to NULL.}

\item{RT1Penalty}{Penalty used for first retention time errors.  Defaults to 1.}
}
\value{
A new metabolite info file with three columns appended to that are the top three Feihn library matches
}
\description{
Takes a metabolite info table from the ConsensusAlign output and annotates with the Feihn library
}
