% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_FAME_Standards.R
\name{Find_FAME_Standards}
\alias{Find_FAME_Standards}
\title{Scans over a list of input files and annotates FAME standards}
\usage{
Find_FAME_Standards(inputFileList, FAME_Frame = system.file("extdata",
  "FIND_FAME_FRAME.txt", package = "CooperR2DGC"), numCores = 4,
  RT1Penalty = 1, RT2Penalty = 10, similarityCutoffWarningThreshold = 85)
}
\arguments{
\item{inputFileList}{Vector of file paths to scan for FAME standards}

\item{numCores}{The number of cores to use for parallel processing. Defaults to 1}

\item{RT1Penalty}{Penalty used for first retention time errors.  Defaults to 1.}

\item{RT2Penalty}{Penalty used for first retention time errors.  Defaults to 10.}

\item{similarityCutoffWarningThreshold}{Similarity score threshold at which to print a warning for user to double check FAME peak for a sample. Defaults to 85.}

\item{FAME_frame}{A parsed FAME standard reference file (find at system.file("extdata", "FIND_FAME_FRAME.txt", package="CooperR2DGC"))}
}
\value{
Will write out new files with properly annotated FAME standards with the extention _FAME_appended.txt added to the end of the file path
}
\description{
Scans over a list of input files and annotates FAME standards
}
